% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spM.R
\name{blocknngp.inla}
\alias{blocknngp.inla}
\title{Spatial multilevel model with block-NNGP prior}
\usage{
blocknngp.inla(
  formula,
  preprocess.data,
  family = "gaussian",
  E = rep(1, dim(preprocess.data$data)[1]),
  offset = NULL,
  phi.prior = c(1, 30),
  sigma2.prior = c(2, 1),
  mc.cores = 1,
  control.compute = list(dic = TRUE, waic = TRUE, config = TRUE),
  control.predictor = list(),
  control.family = list(),
  control.inla = list(),
  control.results = list(),
  control.fixed = list()
)
}
\arguments{
\item{formula}{Model formula, e.g. y~x.}

\item{preprocess.data}{Preprocessed data from \code{\link{preprocess.blocknngp}}.}

\item{family}{A string indicating the likelihood family. The default is gaussian with identity link.
See names(inla.models()$likelihood) for a list of possible alternatives.}

\item{E}{Known component in the mean for the Poisson likelihoods defined as Eexp(lambda)}

\item{offset}{This argument is used to specify an a-priori known and fixed component to be included in the linear predictor during fitting.}

\item{phi.prior}{2*1 vector c(a,b) indicates the uniform prior U(a, b) for spatial decay parameter phi.}

\item{sigma2.prior}{2*1 vector c(a,b) indicates the Inverse-Gamma prior IG(a, b) for spatial marginal variance parameter sigma^2.}

\item{mc.cores}{The number of CPU cores should be used. For Windows user, only 1 core is allowed.}

\item{control...}{See \code{\link[INLA]{inla}}.}
}
\value{
Function that returns a list
\itemize{
\item res - A fitted INLA model, see \code{\link[INLA]{inla}}.
\item Summary statistics of hyperparameters.
}
}
\description{
Fit the spatial multilevel model with block-NNGP prior to the preprocessed data from \code{\link{preprocess.blocknngp}}.
}
