% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normal.R
\name{document_generator_Normal}
\alias{document_generator_Normal}
\title{Simulate data from given parameters.}
\usage{
document_generator_Normal(a, rho, Ts, Lambda, Tau, N, w = NULL, seed = NULL)
}
\arguments{
\item{a:}{dispersion parameter for membership}

\item{rho:}{weight parameter for membership}
}
\value{
A list res containing: 1) X: the document; 2) Y: the labels; 3) G: the untransformed memberships and 4) U: the transformed memberships.
}
\description{
Simulate data from given parameters.
}
