% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract.age.size.r
\name{extract.age_size}
\alias{extract.age_size}
\title{Utility extract function}
\usage{
extract.age_size(lines)
}
\description{
Utility extract function
}
\author{
Alistair Dunn
}

