% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract.ageingerror.R
\name{extract.ageing_error}
\alias{extract.ageing_error}
\title{Utility extract function}
\usage{
extract.ageing_error(lines)
}
\description{
Utility extract function
}
\author{
Alistair Dunn
}

