% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.ageingerror.R
\name{extract.ageingerror}
\alias{extract.ageingerror}
\title{Utility extract function}
\usage{
extract.ageingerror(lines)
}
\description{
Utility extract function
}
\author{
Alistair Dunn
}

