% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract.age.size.r
\name{extract.agesize}
\alias{extract.agesize}
\title{Utility extract function}
\usage{
extract.agesize(lines)
}
\description{
Utility extract function
}
\author{
Alistair Dunn
}

