% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.age.weight.r
\name{extract.ageweight}
\alias{extract.ageweight}
\title{Utility extract function}
\usage{
extract.ageweight(lines)
}
\description{
Utility extract function
}
\author{
Alistair Dunn
}

