% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.derivedquantity.R
\name{extract.derivedquantity}
\alias{extract.derivedquantity}
\title{Helper function for 'extract'}
\usage{
extract.derivedquantity(lines)
}
\arguments{
\item{lines}{list of lines to process}
}
\description{
Helper function for 'extract'
}
\author{
Alistair Dunn
}
