% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract.estimatesummary.R
\name{extract.estimatesummary}
\alias{extract.estimatesummary}
\title{Utility extract function}
\usage{
extract.estimatesummary(lines)
}
\description{
Utility extract function
}
\author{
Alistair Dunn
}

