% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.estimatevalue.R
\name{extract.estimatevalue}
\alias{extract.estimatevalue}
\title{Helper function for 'extract'}
\usage{
extract.estimatevalue(lines)
}
\arguments{
\item{lines}{list of lines to process}
}
\description{
Helper function for 'extract'
}
\author{
Alistair Dunn
}
