% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.layer.R
\name{extract.layer}
\alias{extract.layer}
\title{Helper function for 'extract'}
\usage{
extract.layer(lines)
}
\arguments{
\item{lines}{list of lines to process}
}
\description{
Helper function for 'extract'
}
\author{
Alistair Dunn
}
