% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.MCMC.r
\name{extract.MCMC}
\alias{extract.MCMC}
\title{Helper function for 'extract'}
\usage{
extract.MCMC(lines)
}
\arguments{
\item{lines}{list of lines to process}
}
\description{
Helper function for 'extract'
}
\author{
Alistair Dunn
}
