% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract.observation.R
\name{extract.observation}
\alias{extract.observation}
\title{Utility extract function}
\usage{
extract.observation(lines)
}
\description{
Utility extract function
}
\author{
Alistair Dunn
}

