% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract.partition.biomass.R
\name{extract.partition.biomass}
\alias{extract.partition.biomass}
\title{Utility extract function}
\usage{
extract.partition.biomass(lines)
}
\description{
Utility extract function
}
\author{
Alistair Dunn
}

