% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.partition.biomass.R
\name{extract.partitionbiomass}
\alias{extract.partitionbiomass}
\title{Helper function for 'extract'}
\usage{
extract.partitionbiomass(lines)
}
\arguments{
\item{lines}{list of lines to process}
}
\description{
Helper function for 'extract'
}
\author{
Alistair Dunn
}
