% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.process.R
\name{extract.process}
\alias{extract.process}
\title{Helper function for 'extract'}
\usage{
extract.process(lines)
}
\arguments{
\item{lines}{list of lines to process}
}
\description{
Helper function for 'extract'
}
\author{
Alistair Dunn
}
