% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extract}
\alias{extract}
\title{Generate a multivariate uniform distribution based on the bounds for the extimated parameters}
\usage{
extract(file, path = "", ignore.unknown = FALSE)
}
\arguments{
\item{file}{the name of the input file containing model output to extract}

\item{path}{Optionally, the path to the file}

\item{ignore.unknown}{Ignore unknown objects in the output file}
}
\description{
Generate a multivariate uniform distribution based on the bounds for the extimated parameters
}
\author{
Alistair Dunn
}

