% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.simulations.r
\name{extract.simulations}
\alias{extract.simulations}
\title{Helper function for 'extract'}
\usage{
extract.simulations(file, path = "")
}
\arguments{
\item{lines}{list of lines to process}
}
\description{
Helper function for 'extract'
}
\author{
Alistair Dunn
}
