% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract.size.weight.r
\name{extract.size_weight}
\alias{extract.size_weight}
\title{Utility extract function}
\usage{
extract.size_weight(lines)
}
\description{
Utility extract function
}
\author{
Alistair Dunn
}

