% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.size.weight.r
\name{extract.sizeweight}
\alias{extract.sizeweight}
\title{Helper function for 'extract'}
\usage{
extract.sizeweight(lines)
}
\arguments{
\item{lines}{list of lines to process}
}
\description{
Helper function for 'extract'
}
\author{
Alistair Dunn
}
