% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract.spatial.map.R
\name{extract.spatial.map}
\alias{extract.spatial.map}
\title{Utility extract function}
\usage{
extract.spatial.map(lines)
}
\description{
Utility extract function
}
\author{
Alistair Dunn
}

