% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.spatial.map.R
\name{extract.spatialmap}
\alias{extract.spatialmap}
\title{Helper function for 'extract'}
\usage{
extract.spatialmap(lines)
}
\arguments{
\item{lines}{list of lines to process}
}
\description{
Helper function for 'extract'
}
\author{
Alistair Dunn
}
