% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generate.MVU}
\alias{generate.MVU}
\title{Generate a multivariate uniform distribution based on the bounds for the extimated parameters}
\usage{
generate.MVU(file, path = "", output.file, sample.size = 1)
}
\arguments{
\item{file}{the name of the input file containing the estimated fits}

\item{path}{Optionally, the path to the file}

\item{output.file}{The name of the output file to write randomly generated values}

\item{sample.size}{The number f samples to generate}
}
\description{
Generate a multivariate uniform distribution based on the bounds for the extimated parameters
}
\author{
Sophie Mormede
}

