% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spM.R
\name{get.new.spatial}
\alias{get.new.spatial}
\title{Predict spatial effect at new location}
\usage{
get.new.spatial(
  n.sample = 1,
  new.coord,
  result,
  preprocess.data,
  model = "nngp",
  phi.prior = c(1, 30),
  mc.cores = 1
)
}
\arguments{
\item{n.sample}{Number of posterior samples needed.}

\item{new.coord}{A L*2 matrix of new coordinates whose spatial effect needs to be predicted.}

\item{preprocess.data}{Preprocessed data from \code{\link{preprocess.nngp}} or \code{\link{preprocess.blocknngp}}.}

\item{model}{"nngp" or "blocknngp"; default is "nngp", anything other than "nngp' would be considered as "blocknngp".}

\item{phi.prior}{2*1 vector c(a,b) indicates the uniform prior U(a, b) for spatial decay parameter phi.}

\item{mc.cores}{The number of CPU cores should be used. For Windows user, only 1 core is allowed.}

\item{res}{Result from \code{\link{nngp.inla}} or \code{\link{blocknngp.inla}}.}
}
\value{
Function that returns posterior samples.
\itemize{
\item res - n.sample*L matrix that i_th column contains n.sample posterior samples of i_th location in new.coord.
}
}
\description{
Draw posterior predictive samples for spatial effect at new locations using fitted NNGP or block-NNGP model. Only new locations are allowed,
any location existed in data may cause error.
}
