% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normal.R
\name{get_parameters_Normal}
\alias{get_parameters_Normal}
\title{Calculate the generated distribution given the membership and the topics.}
\usage{
get_parameters_Normal(X, Y, G, Ts, Lambda, Mu)
}
\value{
A list of the parameters of the distributions of data given its memberships.
}
\description{
Calculate the generated distribution given the membership and the topics.
}
