% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm.impute.R
\name{getPrevY.discr}
\alias{getPrevY.discr}
\title{An internal function to compute previous value of
physiological variable Y based on 
discrete-time model (Akushevich et. al., 2005)}
\usage{
getPrevY.discr(y2, u, R, Sigma)
}
\arguments{
\item{y2}{y2}

\item{u}{u (see Akushevich et. al, 2005)}

\item{R}{R (see Akushevich et. al, 2005)}

\item{Sigma}{Sigma (see Akushevich et. al, 2005)}
}
\value{
y1 Previous value of y
}
\description{
An internal function to compute previous value of
physiological variable Y based on 
discrete-time model (Akushevich et. al., 2005)
}

