% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_time-dependent.R
\name{optimize}
\alias{optimize}
\title{A sub-finction that corresponds to the spm_time_dep(...)}
\usage{
optimize(data, starting_params, formulas, verbose, lb, ub, algorithm,
  stopifbound, maxeval = maxeval)
}
\arguments{
\item{data}{a data matrix.}

\item{starting_params}{initial point.}

\item{formulas}{a set of formulas for model parameters.}

\item{verbose}{a verbosing output.}

\item{lb}{lower bound.}

\item{ub}{upper bound.}

\item{algorithm}{some optimization algorithm, for example, NLOPT_LN_NEWUOA.}

\item{stopifbound}{if TRUE then algorithm stops when at least one parameter.}

\item{maxeval}{maximum number of evaluations of the algorithm.
achieved boundaries.}
}
\description{
A sub-finction that corresponds to the spm_time_dep(...)
}

