% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PF.R
\name{PF}
\alias{PF}
\title{Evaluate a preference function}
\usage{
PF(type = "none", x, alpha, ..., rescale = T)
}
\arguments{
\item{type}{A type of preference function. Valid values include constant", "doublenormal", "exponential", "inverselogistic", "logistic", "normal", and "threshold"}

\item{x}{x-values over which to evaluate. Default = 1}

\item{alpha}{The value of the alpha parameter. Default = 1}

\item{...}{the parameters of the preference functions}

\item{rescale}{Rescale the function to have value 1. Default = TRUE}
}
\description{
Evaluate a preference function
}
\author{
Alistair Dunn
}

