% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PFconstant}
\alias{PFconstant}
\title{Evaluate the constant preference function}
\usage{
PFconstant(x, alpha, rescale = T)
}
\arguments{
\item{x}{x-values over which to evaluate. Default = 1}

\item{alpha}{The value of the alpha parameter. Default = 1}

\item{rescale}{Rescale the function to have value 1?. Default = TRUE}
}
\description{
Evaluate the constant preference function
}
\author{
Alistair Dunn
}

