% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFdoublenormal.R
\name{PFdoublenormal}
\alias{PFdoublenormal}
\title{Evaluate the double normal preference function}
\usage{
PFdoublenormal(x, alpha, mu, sigmaL, sigmaR, rescale = T)
}
\arguments{
\item{x}{x-values over which to evaluate. Default = 1}

\item{alpha}{The value of the alpha parameter. Default = 1}

\item{mu}{the mean of the double normal}

\item{sigmaL}{the standard deviation of the left-hand side of the double normal}

\item{sigmaR}{the standard deviation of the right-hand side of the double normal}

\item{rescale}{Rescale the function to have value 1?. Default = TRUE}
}
\description{
Evaluate the double normal preference function
}
\author{
Alistair Dunn
}

