% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFexponential.R
\name{PFexponential}
\alias{PFexponential}
\title{Evaluate the exponential preference function}
\usage{
PFexponential(x, alpha, lambda, rescale = T)
}
\arguments{
\item{x}{x-values over which to evaluate. Default = 1}

\item{alpha}{The value of the alpha parameter. Default = 1}

\item{lambda}{the rate of the exponential}

\item{rescale}{Rescale the function to have value 1?. Default = TRUE}
}
\description{
Evaluate the exponential preference function
}
\author{
Alistair Dunn
}

