% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PFinverselogistic}
\alias{PFinverselogistic}
\title{Evaluate the inverse-logistic preference function}
\usage{
PFinverselogistic(x, alpha, a50, ato95, rescale = T)
}
\arguments{
\item{x}{x-values over which to evaluate. Default = 1}

\item{alpha}{The value of the alpha parameter. Default = 1}

\item{a50}{the a50 value of the inverse logisitic}

\item{ato95}{the ato95 value of the inverse logisitic}

\item{rescale}{Rescale the function to have value 1?. Default = TRUE}
}
\description{
Evaluate the inverse-logistic preference function
}
\author{
Alistair Dunn
}

