% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PFlogistic.R
\name{PFlogistic}
\alias{PFlogistic}
\title{Evaluate the logistic preference function}
\usage{
PFlogistic(x, alpha, a50, ato95, rescale = T)
}
\arguments{
\item{x}{x-values over which to evaluate. Default = 1}

\item{alpha}{The value of the alpha parameter. Default = 1}

\item{a50}{the a50 value of the logisitic}

\item{ato95}{the ato95 value of the logisitic}

\item{rescale}{Rescale the function to have value 1?. Default = TRUE}
}
\description{
Evaluate the logistic preference function
}
\author{
Alistair Dunn
}

