% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PFnormal.R
\name{PFnormal}
\alias{PFnormal}
\title{Evaluate the normal preference function}
\usage{
PFnormal(x, alpha, mu, sigma, rescale = T)
}
\arguments{
\item{x}{x-values over which to evaluate. Default = 1}

\item{alpha}{The value of the alpha parameter. Default = 1}

\item{mu}{the mean of the normal}

\item{sigma}{the standard deviation of the normal}

\item{rescale}{Rescale the function to have value 1?. Default = TRUE}
}
\description{
Evaluate the normal preference function
}
\author{
Alistair Dunn
}

