% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.default}
\alias{predict.default}
\title{Default prediction function for spm}
\usage{
\method{predict}{default}(object, x, dt = 8)
}
\arguments{
\item{object}{An object of class \code{spm}}

\item{x}{A data frame with current values of \code{Y} and \code{t}}

\item{dt}{A time interval. Default: 8.}
}
\value{
A data frame consisting of predicted values of \code{Y},

hazard \code{H}, \code{t}, \code{t.next}.
}
\description{
Default prediction function for spm
}
