% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.spm}
\alias{predict.spm}
\title{Predict function for spm}
\usage{
\method{predict}{spm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{spm}}

\item{data}{A data frame with current values of \code{Y} and \code{t}}

\item{dt}{A time interval. Default: 8.}
}
\value{
A data frame consisting of predicted values of \code{Y},

hazard \code{H}, \code{t}, \code{t.next}.
}
\description{
Predict function for spm
}
\examples{
\dontrun{
library(stpm)
data <- simdata_discr(N=100, format="long")
res <- spm_discrete(data)
splitted <- split(data, data$id)
df <- data.frame()
lapply(1:100, function(i) {df<<-rbind(df,splitted[[i]][dim(splitted[[i]])[1],c("id", "xi", "t1", "y1")])})
names(df) <- c("id", "xi", "t", "y")
predicted <- predict(object=res, data=df, dt=3)
head(predicted)
}

}
