% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Data pre-processing for analysis with stochastic process model methodology.}
\usage{
prepare_data(longdat, vitstat, interval = 1, col.status = "IsDead",
  col.id = "ID", col.age = "Age", col.age.event = "LSmort",
  covariates = c("DBP", "BMI", "DBP1", "DBP2", "Weight", "Height"),
  verbose = TRUE)
}
\arguments{
\item{longdat}{A table with longitude records.}

\item{vitstat}{A table with vital statistics (mortality).}

\item{interval}{A number of breaks between observations for discrete simulation. Default = 1 (no breaks).}

\item{col.status}{A name of column containing status variable (0/1 which indicate alive/dead).}

\item{col.id}{A name of column containing patient ID. This ID should be the same in both longdat and vitstat tables.}

\item{col.age}{A name of age column.}

\item{col.age.event}{A name of event column.}

\item{covariates}{A list of covariates.}

\item{verbose}{A verbosing output indicator. Default=TRUE.}
}
\value{
A list of two elements: first element contains a data table for continuous case, with arbitrary intervals between observations  and 
second element contains a data table for a discrete case (fixed intervals between observations).
}
\description{
Data pre-processing for analysis with stochastic process model methodology.
}
\examples{
library(spm)
#Reading longitude data:
ldat <- read.csv(system.file("data","longdat.csv",package="spm"))
# Prepare data for optimization:
vdat <- read.csv(system.file("data","vitstat.csv",package="spm"))
data <- prepare_data(longdat=ldat, vitstat=vdat,interval=1, col.status="IsDead", col.id="ID", col.age="Age", col.age.event="LSmort", covariates=c("DBP"), verbose=TRUE)
# Parameters estimation:
pars <- spm(data,k = 1)
pars
}

