% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data_cont}
\alias{prepare_data_cont}
\title{Prepares continuouts-time dataset.}
\usage{
prepare_data_cont(longdat, vitstat, col.status.ind, col.id.ind, col.age.ind,
  col.age.event.ind, col.covar.ind, verbose, impute)
}
\arguments{
\item{longdat}{a longitudinal study dataset.}

\item{vitstat}{vital (mortality) statistics.}

\item{col.status.ind}{index of "status" column.}

\item{col.id.ind}{subject id column index.}

\item{col.age.ind}{index of the age column.}

\item{col.age.event.ind}{an index of the column which represents the time in which event occured.}

\item{col.covar.ind}{a set of column indexes which represent covariates.}

\item{verbose}{turns on/off verbosing output.}

\item{impute}{Multiple imputation ndicator. If TRUE then missing observations will be imputed with multiple imputation.}
}
\description{
Prepares continuouts-time dataset.
}

