% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.spm.config.R
\name{read.spm.config}
\alias{read.spm.config}
\title{Model configuration read function}
\usage{
read.spm.config(file, path = "")
}
\arguments{
\item{file}{the name of the input file containing model configuration}

\item{path}{Optionally, the path to the file}
}
\description{
This function reads an SPM configuration file and returns a list object in R. Where each element is a command and subcommand from the configuration file
}
\author{
Alistair Dunn
}
