% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_continuous.R
\name{setlb}
\alias{setlb}
\title{Sets lower boundaries.}
\usage{
setlb(k, params)
}
\arguments{
\item{k}{number of dimensions.}

\item{params}{a vector of parameter values.}
}
\description{
Sets lower boundaries.
}

