% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_continuous.R
\name{setub}
\alias{setub}
\title{Sets upper boundaries.}
\usage{
setub(k, params)
}
\arguments{
\item{k}{number of dimensions.}

\item{params}{a vector of parameter values.}
}
\description{
Sets upper boundaries.
}

