% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simdataND.R
\name{sim_discrete}
\alias{sim_discrete}
\title{Multi-dimension simulation function
It uses a, f1, Q, f, b, mu0 and theta
as input parameters.}
\usage{
sim_discrete(N = 100, a = -0.05, f1 = 80, Q = 2e-08, f = 80, b = 5,
  mu0 = 1e-05, theta = 0.08, ystart = c(80), tstart = 30, tend = 105,
  dt = 1, k = 1)
}
\arguments{
\item{N}{Number of individuals}

\item{a}{A k by k matrix, which characterize the rate of the adaptive response.}

\item{f1}{A particular state, which if a deviation from the normal (or optimal). This is a vector with length of k.}

\item{Q}{A matrix k by k, which is a non-negative-definite symmetric matrix.}

\item{f}{A vector-function (with length k) of the normal (or optimal) state.}

\item{b}{A diffusion coefficient, k by k matrix.}

\item{mu0}{mortality at start period of time.}

\item{theta}{A displacement coefficient of the Gompertz function.}

\item{ystart}{A vector with length equal to number of dimensions used, defines starting values of covariates.}

\item{tstart}{A number that defines starting time (30 by default).}

\item{tend}{A number, defines final time (105 by default).}

\item{dt}{A time step (1 by default).}

\item{k}{number of dimensions (k = 1 by default).}
}
\value{
A table with simulated data.
}
\description{
Multi-dimension simulation function
It uses a, f1, Q, f, b, mu0 and theta
as input parameters.
}
\examples{
library(spm)
data <- sim(N=1000, ystart=c(75, 94), k=1)
head(data)
}

