% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simdataND.R
\name{simdata}
\alias{simdata}
\title{Function that simulates data using u, R, epsilon, mu0, b, Q, theta}
\usage{
simdata(N = 100, u = 8, R = 0.95, epsilon = 5, mu0 = 2e-05, b = 10,
  Q = 2e-08, theta = 0.08, tstart = 30, ystart = 80, dt = 1,
  tmax = 105, k = 1)
}
\arguments{
\item{N}{Number of individuals}

\item{u}{A drift vector with length of k.}

\item{R}{A k by k regression matrix.}

\item{epsilon}{A time-dependent normally distributed random vector (size=k).}

\item{mu0}{mortality at start period of time.}

\item{b}{A diffusion coefficient, k by k matrix.}

\item{Q}{A matrix k by k, which is a non-negative-definite symmetric matrix.}

\item{theta}{A displacement coefficient of the Gompertz function.}

\item{tstart}{A number that defines starting time (30 by default).}

\item{ystart}{A vector with length equal to number of dimensions used, defines starting values of covariates.}

\item{dt}{A time step (1 by default).}

\item{k}{Number of dimensions (k = 1 by default).}

\item{tend}{A number, defines final time (105 by default).}
}
\value{
A table with simulated data.
}
\description{
Function that simulates data using u, R, epsilon, mu0, b, Q, theta
}
\examples{
library(spm)
data <- simdata(N=1000, ystart=c(75, 94), k=1)
head(data)
}

