% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simdata_cont_MD.R
\name{simdata_cont_MD}
\alias{simdata_cont_MD}
\title{Multi-dimensional simulation function for continuous trait.}
\usage{
simdata_cont_MD(N = 10, aH = -0.05, f1H = 80, QH = 2e-07, fH = 80,
  bH = 5, mu0H = 2e-05, thetaH = 0.08, step = 0.05, tstart = 30,
  tend = 105, ystart = 80, sd0 = 4, k = 1)
}
\arguments{
\item{N}{Number of individuals.}

\item{tstart}{A number that defines starting time (30 by default).}

\item{tend}{A number, defines final time (105 by default).}

\item{ystart}{A vector with length equal to number of dimensions used, defines starting values of covariates.}

\item{k}{number of dimensions (k = 1 by default).}

\item{a}{A k by k matrix, which characterize the rate of the adaptive response.}

\item{f1}{A particular state, which if a deviation from the normal (or optimal). This is a vector with length of k.}

\item{Q}{A matrix k by k, which is a non-negative-definite symmetric matrix.}

\item{f}{A vector-function (with length k) of the normal (or optimal) state.}

\item{b}{A diffusion coefficient, k by k matrix.}

\item{mu0}{mortality at start period of time.}

\item{theta}{A displacement coefficient of the Gompertz function.}
}
\value{
A table with simulated data.
}
\description{
Multi-dimensional simulation function for continuous trait.
}
\examples{
library(spm)
dat <- simdata_cont(N=2500)
dat
}

