% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spM.R
\name{spatial.plot}
\alias{spatial.plot}
\title{Spatial plot}
\usage{
spatial.plot(
  result,
  preprocess.data,
  coord = preprocess.data$unique.coord,
  values = result$res$summary.random$idx$mean,
  nx = 200,
  ny = 200,
  zlim = NULL
)
}
\arguments{
\item{result}{Result from \code{\link{nngp.inla}} or \code{\link{blocknngp.inla}}.}

\item{preprocess.data}{Preprocessed data from \code{\link{preprocess.nngp}} or \code{\link{preprocess.blocknngp}}.}

\item{coord}{K*2 coordinates matrix. Specify for your customized coordinates, do not need to specify when you want estimated spatial surface of fitted data.}

\item{values}{1*k vector of values at corresponding coordinates. Specify for your customized values, do not need to specify when you want estimated spatial surface of fitted data.}

\item{nx}{Dimension of output grid in x direction.}

\item{ny}{Dimension of output grid in x direction.}

\item{zlim}{Adjust the ranges of values.}
}
\value{
Function that returns a figure.
}
\description{
Create spatial interpolation plot. Specify result and preprocess.data to get estimated spatial surface of fitted data.
}
