% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spm.extract.simulated.observation.r
\name{spm.extract.simulated.observation}
\alias{spm.extract.simulated.observation}
\title{utility function}
\usage{
spm.extract.simulated.observation(lines)
}
\description{
utility function
}
\author{
Alistair Dunn
}

