% Generated by roxygen2 (4.0.2): do not edit by hand
\name{spm.get.lines}
\alias{spm.get.lines}
\title{utility function}
\usage{
spm.get.lines(lines, from = -1, to = -1, contains = "",
  starts.with = "", clip.to = "", clip.from = "", clip.to.match = "",
  clip.from.match = "", ...)
}
\description{
utility function
}
\author{
Alistair Dunn
}

