% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spm.is.whole.number.R
\name{spm.is.whole.number}
\alias{spm.is.whole.number}
\title{utility function}
\usage{
spm.is.whole.number(string)
}
\description{
utility function
}
\author{
Alistair Dunn
}

