% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spm.make.filename.R
\name{spm.make.filename}
\alias{spm.make.filename}
\title{utility function}
\usage{
spm.make.filename(file, path = "")
}
\description{
utility function
}
\author{
Alistair Dunn
}

