% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm.R
\name{spm}
\alias{spm}
\title{Stochastic Process Modelling (SPM)
A main function that estimates parameters a, f1, Q, f, b, mu0, theta
from given dataset.}
\usage{
spm(dat, k = 2, verbose = F, tol = NULL)
}
\arguments{
\item{dat}{A dataset.}

\item{k}{Number of dimensions.}

\item{verbose}{A verbosing output indicator.}

\item{tol}{A tolerance threshold for matrix inversion.}
}
\value{
A list of (1) Estimated starting point (from quick discrete optimization) and 
(2) Estimated coefficients.
}
\description{
Stochastic Process Modelling (SPM)
A main function that estimates parameters a, f1, Q, f, b, mu0, theta
from given dataset.
}
\examples{
library(spm)
#Prepare data for optimization
longdat <- read.csv(system.file("data","longdat.csv",package="spm"))
vitstat <- read.csv(system.file("data","vitstat.csv",package="spm"))
data=prepare_data(longdat=longdat, vitstat=vitstat,interval=1, col.status="IsDead", col.id="ID", col.age="Age", col.age.event="LSmort", covariates=c("DBP"), verbose=T)
#Parameters estimation:
pars=spm(data,k = 1)
pars
}

