% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm.recodevector.R
\name{spm.recodevector}
\alias{spm.recodevector}
\title{utility function}
\usage{
spm.recodevector(in.data, from.vals, to.vals)
}
\description{
utility function
}
\author{
Alistair Dunn
}

