% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm.remove.first.words.R
\name{spm.remove.first.words}
\alias{spm.remove.first.words}
\title{utility function}
\usage{
spm.remove.first.words(string, words = 1)
}
\description{
utility function
}
\author{
Alistair Dunn
}

