% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spm.string.to.vector.of.numbers.R
\name{spm.string.to.vector.of.numbers}
\alias{spm.string.to.vector.of.numbers}
\title{utility function}
\usage{
spm.string.to.vector.of.numbers(string, sep = " ")
}
\description{
utility function
}
\author{
Alistair Dunn
}

