% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spm.string.to.vector.of.words.R
\name{spm.string.to.vector.of.words}
\alias{spm.string.to.vector.of.words}
\title{utility function}
\usage{
spm.string.to.vector.of.words(string, sep = " ")
}
\description{
utility function
}
\author{
Alistair Dunn
}

