% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_continuous_1D.R
\name{spm_continuous_1D}
\alias{spm_continuous_1D}
\title{Continuous one-dimensional optimization}
\usage{
spm_continuous_1D(dat, a = -0.05, f1 = 80, Q = 2e-08, f = 80, b = 5,
  mu0 = 2e-05, theta = 0.08)
}
\arguments{
\item{dat}{A data table.}

\item{a}{A starting value of the rate of adaptive response to any deviation of Y from f1(t).}

\item{f1}{A starting value of the average age trajectories of the variables which process is forced to follow.}

\item{Q}{Starting values of the quadratic hazard term.}

\item{f}{A starting value of the "optimal" value of variable which corresponds to the minimum of hazard rate at a respective time.}

\item{b}{A starting value of a diffusion coefficient representing a strength of the random disturbance from Wiener Process.}

\item{mu0}{A starting value of the baseline hazard.}

\item{theta}{A starting value of the parameter theta (axe displacement of Gompertz function).}
}
\value{
A set of estimated parameters a, f1, Q, f, b, mu0, theta.
}
\description{
Continuous one-dimensional optimization
}
\details{
\code{spm_integral_1D} runs much slower that discrete but more precise and can handle time intervals with different lengths.
}

