% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_discrete_MD.R
\name{spm_discrete_MD}
\alias{spm_discrete_MD}
\title{Discrete multi-dimensional optimization}
\usage{
spm_discrete_MD(dat, k = 1, theta_range = seq(0.001, 0.09, by = 0.001),
  tol = NULL)
}
\arguments{
\item{dat}{A data table.}

\item{k}{A number of dimensions.}

\item{theta_range}{A range of theta parameter (axe displacement of Gompertz function), default: from 0.001 to 0.09 with step of 0.001.}

\item{tol}{A tolerance threshold for matrix inversion (NULL by default).}
}
\value{
A list of two elements: (1) estimated parameters u, R, b, epsilon, Q, mu0, theta and
(2) estimated parameters a, f1, Q, f, b, mu0, theta. Note: b and mu0 from first list are different 
from b and mu0 from the second list.
}
\description{
Discrete multi-dimensional optimization
}
\details{
This function is way much faster that continuous \code{spm_continuous_MD(...)} (but less precise) and used mainly in 
estimation a starting point for the \code{spm_continuous_MD(...)}.
}
\examples{
library(spm)
# Reading longitudinal data
longdat <- read.csv(system.file("data","longdat.csv",package="spm"))
# Prepare data for optimization
vitstat <- read.csv(system.file("data","vitstat.csv",package="spm"))
data <- prepare_data(longdat=longdat, vitstat=vitstat,interval=1, col.status="IsDead", col.id="ID", col.age="Age", col.age.event="LSmort", covariates=c("DBP"), verbose=T)
# Parameters estimation
pars <- spm_discrete_MD(data[[2]], k=1, theta_range=seq(0.001,0.09,by=0.001), tol=NULL)
pars
}

