% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_integral_MD.R
\name{spm_integral_MD}
\alias{spm_integral_MD}
\title{Continuous multi-dimensional optimization
It is much slower that discrete but more precise and can handle time intervals with different lengths.}
\usage{
spm_integral_MD(dat, parameters, k, verbose = F)
}
\arguments{
\item{dat}{A data table.}

\item{parameters}{A starting pont (a vector).}

\item{k}{A number of dimensions.}

\item{verbose}{An indicator of verbosing output.}

\item{tol}{A tolerance threshold for matrix inversion.}
}
\value{
A list of two elements: (1) parameters a, f1, Q, f, b, mu0, theta; (2) An output from "optim" function used 
for maximum likelihood estimation.
}
\description{
Continuous multi-dimensional optimization
It is much slower that discrete but more precise and can handle time intervals with different lengths.
}
\examples{
#'library(spm)
# Reading longitude data:
longdat <- read.csv(system.file("data","longdat.csv",package="spm"))
# Prepare data for optimization:
vitstat <- read.csv(system.file("data","vitstat.csv",package="spm"))
# Remove unneeded NAs:
longdat.nonan <- longdat[which(is.na(longdat$Age) == F),]
vitstat.nonan <- vitstat[which(is.na(vitstat$BirthCohort) == F),]
dat=prepare_data(longdat=longdat.nonan, vitstat=vitstat.nonan,interval=1, col.status="IsDead", col.id="ID", col.age="Age", col.age.next="AgeNext", col.age.event="LSmort", covariates=c("DBP"), verbose=T)
# Parameters estimation:
dat<-[,1:6]
pars=spm_integral_MD(dat, parameters=c(-0.05, 80, 2e-8, 80, 5, 2e-5, 0.08), k = 1)
pars
}

