% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normal.R
\name{SPM_predicting_Normal}
\alias{SPM_predicting_Normal}
\title{Given the SPM model, predict labels for a new document.}
\usage{
SPM_predicting_Normal(X, Lambda, Mu, a, rho, Ts, nsample, seed, w = NULL)
}
\value{
A list containing the posterior distribution of labels and the estimated labels.
}
\description{
Given the SPM model, predict labels for a new document.
}
