% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normal.R
\name{SPM_simulator_Normal}
\alias{SPM_simulator_Normal}
\title{Given the estimated topics and memberships, simulate data}
\usage{
SPM_simulator_Normal(
  Lambda = NULL,
  Tau = NULL,
  Mu = NULL,
  S = NULL,
  Ts = NULL,
  G = NULL,
  Y = NULL,
  U = NULL,
  seed = NULL
)
}
\value{
X: the simulated data
}
\description{
Given the estimated topics and memberships, simulate data
}
