\name{sw}
\alias{sw}
\docType{data}
\title{A dataset of grids for producing spatial predictions of seabed mud content in the southwest Australia Exclusive Economic Zone}
\description{This dataset contains 500703 rows of 2 variables including
longitude (long), latitude (lat).}
\usage{data("sw")}
\format{
  A data frame with 500703 rows on the following 2 variables.
  \describe{
    \item{\code{long}}{a numeric vector, decimal degree}
    \item{\code{lat}}{a numeric vector, decimal degree}
  }
}
\details{For details, please check the source.}
\source{Li, J., Potter, A., Huang, Z., Daniell, J.J., Heap, A., 2010. Predicting Seabed Mud Content across the Australian Margin: Comparison of Statistical and Mathematical Techniques Using a Simulation Experiment. Geoscience Australia, 2010/11, 146pp.}
\keyword{datasets}
