\name{swmud}
\alias{swmud}
\docType{data}
\title{A dataset of seabed mud content in the southwest Australia Exclusive Economic Zone}
\description{This dataset contains 177 samples of 3 variables including
longitude (long), latitude (lat), mud content (mud).}
\usage{data("swmud")}
\format{
  A data frame with 177 observations on the following 3 variables.
  \describe{
    \item{\code{long}}{a numeric vector, decimal degree}
    \item{\code{lat}}{a numeric vector, decimal degree}
    \item{\code{mud}}{a numeric vector, percentage}
  }
}
\details{For details, please check the source.}
\source{Li, J., Potter, A., Huang, Z., Daniell, J.J., Heap, A., 2010. Predicting Seabed Mud Content across the Australian Margin: Comparison of Statistical and Mathematical Techniques Using a Simulation Experiment. Geoscience Australia, 2010/11, 146pp.}
\keyword{datasets}
