% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.R
\name{T_generator}
\alias{T_generator}
\title{Given the model setting, generate the transformation matrices.}
\usage{
T_generator(k0, k1, nlabel)
}
\description{
Given the model setting, generate the transformation matrices.
}
