% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.R
\name{trans_to_matrix}
\alias{trans_to_matrix}
\title{Helpful function to summarize the sampling results.}
\usage{
trans_to_matrix(x, d1, d2, byfeature = TRUE)
}
\description{
Helpful function to summarize the sampling results.
}
