% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.SPM.config.R
\name{write.spm.config}
\alias{write.spm.config}
\title{Model configuration read function}
\usage{
write.spm.config(object, file, path = "")
}
\arguments{
\item{object}{the name of the object containing model configuration}

\item{file}{The file to write to}

\item{path}{Optionally, the path to the file}
}
\description{
This function writes an SPM configuration file. Where each element is a command and subcommand from the configuration file
}
\author{
Alistair Dunn
}
