% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbar-uu-r.R
\name{xbaruur}
\alias{xbaruur}
\title{X-bar chart for unknown mean and unknown standard deviation (Range Based) of process data}
\usage{
xbaruur(x, k, n)
}
\arguments{
\item{x}{is a data frame of values representing process data for subgroup mean (1) and range (2)}

\item{k}{is the sigma limits for the control chart}

\item{n}{is the number of observations in each subgroup}
}
\value{
A X-bar control chart
}
\description{
X-bar chart for unknown mean and unknown standard deviation (Range Based) of process data
}
\examples{
xbaruur(warpbreaks$breaks, 3, 5)
}

