% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbar-uu-s.R
\name{xbaruus}
\alias{xbaruus}
\title{X-bar chart for unknown mean and unknown standard deviation (s based) of process data}
\usage{
xbaruus(x, k, n)
}
\arguments{
\item{x}{is a data frame of values representing process data for subgroup mean (1) and standard deviation (2)}

\item{k}{is the sigma limits for the control chart}

\item{n}{is the number of observations in each subgroup}
}
\value{
A X-bar control chart
}
\description{
X-bar chart for unknown mean and unknown standard deviation (s based) of process data
}
\examples{
xbaruus(warpbreaks$breaks, 3, 5)
}

