% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normal.R
\name{BPM_membership_Normal}
\alias{BPM_membership_Normal}
\title{Given the BPM model, estimate the membership for new data points.}
\usage{
BPM_membership_Normal(X, Lambda, Mu, a, rho, ntrace, nchain, nskip)
}
\value{
A matrix of the estimated memberships.
}
\description{
Given the BPM model, estimate the membership for new data points.
}
