% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calc.abundance}
\alias{calc.abundance}
\title{Generate a multivariate uniform distribution based on the bounds for the extimated parameters}
\usage{
calc.abundance(data, categories = NULL, ages = NULL, total = T)
}
\arguments{
\item{data}{data representing a partition}

\item{categories}{Optional subset of categories}

\item{ages}{Optional subset of ages}

\item{total}{If true, then sum over selected categories and ages. Default = TRUE}
}
\description{
Generate a multivariate uniform distribution based on the bounds for the extimated parameters
}
\author{
Alistair Dunn
}

