% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract.covariance.R
\name{extract.covariance}
\alias{extract.covariance}
\title{Utility extract function}
\usage{
extract.covariance(lines)
}
\description{
Utility extract function
}
\author{
Alistair Dunn
}

