% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract.derivedquantity.R
\name{extract.derivedquantity}
\alias{extract.derivedquantity}
\title{Utility extract function}
\usage{
extract.derivedquantity(lines)
}
\description{
Utility extract function
}
\author{
Alistair Dunn
}

