% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract.derivedquantitybycell.R
\name{extract.derivedquantitybycell}
\alias{extract.derivedquantitybycell}
\title{Utility extract function}
\usage{
extract.derivedquantitybycell(lines)
}
\description{
Utility extract function
}
\author{
Alistair Dunn
}

