% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract.estimatevalue.R
\name{extract.estimatevalue}
\alias{extract.estimatevalue}
\title{Utility extract function}
\usage{
extract.estimatevalue(lines)
}
\description{
Utility extract function
}
\author{
Alistair Dunn
}

