% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract.initialisationphase.R
\name{extract.initialisationphase}
\alias{extract.initialisationphase}
\title{Utility extract function}
\usage{
extract.initialisationphase(lines)
}
\arguments{
\item{lines}{a file of scanned lines}
}
\description{
Utility extract function
}
\author{
Alistair Dunn
}

