% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.layerderivedview.R
\name{extract.layerderivedview}
\alias{extract.layerderivedview}
\title{Utility extract function}
\usage{
extract.layerderivedview(lines)
}
\description{
Utility extract function
}
\author{
Alistair Dunn
}

