% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.objectivefunction.R
\name{extract.objectivefunction}
\alias{extract.objectivefunction}
\title{Utility extract function}
\usage{
extract.objectivefunction(lines)
}
\description{
Utility extract function
}
\author{
Alistair Dunn
}

