% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.partition.biomass.R
\name{extract.partitionbiomass}
\alias{extract.partitionbiomass}
\title{Utility extract function}
\usage{
extract.partitionbiomass(lines)
}
\description{
Utility extract function
}
\author{
Alistair Dunn
}

