% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract.randomnumberseed.R
\name{extract.randomnumberseed}
\alias{extract.randomnumberseed}
\title{Utility extract function}
\usage{
extract.randomnumberseed(lines)
}
\description{
Utility extract function
}
\author{
Alistair Dunn
}

