% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract.selectivity.R
\name{extract.selectivity}
\alias{extract.selectivity}
\title{Utility extract function}
\usage{
extract.selectivity(lines)
}
\description{
Utility extract function
}
\author{
Alistair Dunn
}

