% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.simulations.r
\name{extract.simulations}
\alias{extract.simulations}
\title{Utility extract function}
\usage{
extract.simulations(file, path = "")
}
\description{
Utility extract function
}
\author{
Alistair Dunn
}

