% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.size.weight.r
\name{extract.sizeweight}
\alias{extract.sizeweight}
\title{Utility extract function}
\usage{
extract.sizeweight(lines)
}
\description{
Utility extract function
}
\author{
Alistair Dunn
}

