% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PFthreshold.R
\name{PFthreshold}
\alias{PFthreshold}
\title{Evaluate the threshold preference function}
\usage{
PFthreshold(x, alpha, N, lambda, rescale = T)
}
\arguments{
\item{x}{x-values over which to evaluate. Default = 1}

\item{alpha}{The value of the alpha parameter. Default = 1}

\item{N}{the threshhold value}

\item{rescale}{Rescale the function to have value 1?. Default = TRUE}
}
\description{
Evaluate the threshold preference function
}
\author{
Alistair Dunn
}

