% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata_time-dep.R
\name{simdata_time_dep}
\alias{simdata_time_dep}
\title{Simulation function for continuous trait with time-dependant coefficients.}
\usage{
simdata_time_dep(N = 10, formulas = list(at = "-0.05", f1t = "80", Qt =
  "2e-7*exp(0.08*t)", ft = "80", bt = "5", mu0t = "2e-5*exp(0.08*t)"),
  step = 0.05, tstart = 30, tend = 105, ystart = 80, sd0 = 4, k = 1)
}
\arguments{
\item{N}{Number of individuals.}

\item{formulas}{: a list of formulas that define age (time) - dependency. Default: list(at="a", f1t="f1", Qt="Q*exp(theta*t)", ft="f", bt="b", mu0t="mu0*exp(theta*t)")}

\item{tstart}{A number that defines starting time (30 by default).}

\item{tend}{A number, defines final time (105 by default).}

\item{ystart}{A starting value of covariates.}
}
\value{
A table with simulated data.
}
\description{
Simulation function for continuous trait with time-dependant coefficients.
}
\examples{
library(spm)
dat <- simdata_time_dep(N=2500)
dat
}

