\name{spm-package}
\alias{spm-package}
\alias{spm}
\docType{package}
\title{
\packageTitle{spm}
}
\description{
\packageDescription{spm}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{spm}
\packageIndices{spm}
}
\author{
\packageAuthor{spm}

Maintainer: \packageMaintainer{spm}
}
\keyword{ package }
