\name{spm-package}
\alias{spm-package}
\alias{spm}
\docType{package}
\title{
Data Weighting package
}
\description{
A set of R functions for extracting and processing SPM output.
}
\details{
\tabular{ll}{
Package: \tab spm\cr
Date: \tab 2013-02-12\cr
License: \tab See the SPM manual, or use 'spm -l' at the command line to view the SPM license.\cr
URL: \tab http://www.niwa.co.nz\cr
Copyright: \tab National Institute of Water & Atmospheric Research (NIWA).\cr
}

Index:
\preformatted{
extract                 Extract elements from an SPM output file
spm-package             SPM package
}
}
\author{
A. Dunn

Maintainer: A Dunn <a.dunn@niwa.co.nz>
}
\keyword{ package }
