% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm.dplot.R
\name{spm.dplot}
\alias{spm.dplot}
\title{utility function}
\usage{
spm.dplot(..., name = T, quantiles = c(0.5), plot.mean = F, main = "",
  xlab = "", ylab = "", ylim, srtx = 0, bw = "nrd0", adjust = 1/3,
  adj = 0, fill = F, gap = 0.2)
}
\description{
utility function
}
\author{
Alistair Dunn
}

