% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spm.R
\name{spm}
\alias{spm}
\title{Stochastic Process Modelling (SPM)
A main function that estimates parameters a, f1, Q, f, b, mu0, theta
from given dataset.}
\usage{
spm(dat, k = 2, verbose = F, tol = NULL)
}
\arguments{
\item{dat}{A dataset.}

\item{k}{Number of dimensions.}

\item{verbose}{A verbosing output indicator.}

\item{tol}{A tolerance threshold for matrix inversion.}
}
\value{
A list of (1) Estimated starting point (from quick discrete optimization) and
(2) Estimated coefficients.
}
\description{
Stochastic Process Modelling (SPM)
A main function that estimates parameters a, f1, Q, f, b, mu0, theta
from given dataset.
}
\examples{
library(spm)
# Reading longitude data:
longdat <- read.csv(system.file("data","longdat.csv",package="spm"))
# Prepare data for optimization:
vitstat <- read.csv(system.file("data","vitstat.csv",package="spm"))
# Remove unneeded NAs:
longdat.nonan <- longdat[which(is.na(longdat$Age) == F),]
vitstat.nonan <- vitstat[which(is.na(vitstat$BirthCohort) == F),]
data=prepare_data(longdat=longdat.nonan, vitstat=vitstat.nonan,interval=1, col.status="IsDead", col.id="ID", col.age="Age", col.age.next="AgeNext", col.age.event="LSmort", covariates=c("DBP"), verbose=T)
# Parameters estimation:
pars=spm(data,k = 1)
pars
}

