% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm.R
\name{spm}
\alias{spm}
\title{A central function that estimates Stochastic Process Model (SPM) parameters a from given dataset.}
\usage{
spm(x, model = "discrete", formulas = NULL, verbose = FALSE, tol = NULL)
}
\arguments{
\item{x}{A dataset: is the output from prepare_data(...) function and consists of two separate data tables:
(1) a data table for continuous-time model and (2) a data table for discrete-time model.}

\item{model}{A model type. Choices are: "discrete", "continuous" or "time-dependent".}

\item{formulas}{A list of parameter formulas used in the "time-dependent" model.}

\item{verbose}{A verbosing output indicator (FALSE by default).}

\item{tol}{A tolerance threshold for matrix inversion (NULL by default).}
}
\value{
For "discrete" and "continuous" model types: 
(1) a list of model parameter estimates for the discrete model type described in "", Akushevich et al, 2005 and  
(2) a list of model parameter estimates for the continuous model type described in "", Yashin et al, 2007.

For the "time-dependent" model (model parameters depend on time): a set of model parameter estimates.
}
\description{
A central function that estimates Stochastic Process Model (SPM) parameters a from given dataset.
}
\examples{
\dontrun{ 
library(spm)
#Prepare data for optimization
data <- prepare_data(x=system.file("data","longdat.csv",package="spm"), 
				   y=system.file("data","vitstat.csv",package="spm"))
#Parameters estimation (default model: discrete-time):
p.discr.model <- spm(data)
p.discr.model
# Continuous-time model:
p.cont.model <- spm(data, model="continuous")
p.cont.model
#Model with time-dependent coefficients:
data <- prepare_data(x=system.file("data","longdat.csv",package="spm"), 
				   y=system.file("data","vitstat.csv",package="spm"), 
				   covariates="BMI")
p.td.model <- spm(data, model="time-dependent")
p.td.model
}
}

