% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm.R
\name{spm}
\alias{spm}
\title{Stochastic Process Modelling (SPM)
A main function that estimates parameters a, f1, Q, f, b, mu0, theta
from given dataset.}
\usage{
spm(x, model = "discrete", formulas = list(at = "a", f1t = "f1", Qt =
  "Q*exp(theta*t)", ft = "f", bt = "b", mu0t = "mu0*exp(theta*t)"),
  verbose = F, tol = NULL)
}
\arguments{
\item{x}{A dataset.}

\item{model}{A model type: c("continuous", "time-dependent").}

\item{formulas}{A list of formulas used in time-dependent model.}

\item{verbose}{A verbosing output indicator.}

\item{tol}{A tolerance threshold for matrix inversion.}
}
\value{
A list of (1) coefficient estimates from discrete model and 
(if appropriate model type was provided): 
(2) coefficient estimates from continuous model and
(3) coefficient estimates from time-dependent model.
}
\description{
Stochastic Process Modelling (SPM)
A main function that estimates parameters a, f1, Q, f, b, mu0, theta
from given dataset.
}
\examples{
library(spm)
#Prepare data for optimization
longdat <- read.csv(system.file("data","longdat.csv",package="spm"))
vitstat <- read.csv(system.file("data","vitstat.csv",package="spm"))
data=prepare_data(longdat=longdat, vitstat=vitstat,interval=1, col.status="IsDead", col.id="ID", col.age="Age", col.age.event="LSmort", covariates=c("DBP"), verbose=T)
#Parameters estimation:
pars=spm(data)
pars
}

