% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_time-dependent.R
\name{spm_time_dep}
\alias{spm_time_dep}
\title{spm_time_dep : a function that can handle time-dependant coefficients:}
\usage{
spm_time_dep(data, start = list(a1 = -0.5, a2 = 0.2, f1 = 80, Q = 2e-08, f =
  80, b = 5, mu0 = 1e-05, theta = 0.08), formulas = list(at = "a1*t+a2", f1t =
  "f1", Qt = "Q*exp(theta*t)", ft = "f", bt = "b", mu0t = "mu0*exp(theta*t)"),
  verbose = TRUE, lower_bound = NULL, upper_bound = NULL, factr = 1e-16,
  lmult = 0.5, umult = 2)
}
\arguments{
\item{start}{: a list of starting parameters, default: llist(a=-0.5, f1=80, Q=2e-8, f=80, b=5, mu0=1e-5, theta=0.08),}

\item{formulas}{: a list of formulas that define age (time) - dependency. Default: list(at="a1*t+a2", f1t="f1", Qt="Q*exp(theta*t)", ft="f", bt="b", mu0t="mu0*exp(theta*t)")}
}
\value{
optimal coefficients
}
\description{
spm_time_dep : a function that can handle time-dependant coefficients:
}
\examples{
library(spm)
#Data preparation:
N <- 1000
data <- simdata_time_dep(N=2500)
opt.par <- spm_time_dep(data[,2:6], formulas=list(at="a", f1t="f1", Qt="Q*exp(theta*t)", ft="f", bt="b", mu0t="mu0*exp(theta*t)"), start=list(a=-0.5, f1=80, Q=2e-8, f=80, b=5, mu0=1e-5, theta=0.08))
opt.par
}

