% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normal.R
\name{SPM_training_Normal}
\alias{SPM_training_Normal}
\title{Fit a SPM model on the training set given the hyperparameters.}
\usage{
SPM_training_Normal(
  X,
  Y,
  Ts,
  b,
  alpha,
  mu_Mu,
  sigma2_Mu,
  alpha_Lambda,
  beta_Lambda,
  ntrace,
  nchain,
  nskip
)
}
\value{
A list containing all estimated parameters.
}
\description{
Fit a SPM model on the training set given the hyperparameters.
}
