% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.SPM.config.R
\name{write.spm.config}
\alias{write.spm.config}
\title{Model configuration read function}
\usage{
write.spm.config(object, file, path = "", header = "SPM config file",
  date = FALSE)
}
\arguments{
\item{object}{the name of the object containing model configuration}

\item{file}{The file to write to}

\item{path}{Optionally, the path to the file}

\item{header}{Optionally, a text header to append as a comment at the top of the file}

\item{date}{Optionally, a comment giving the current date}
}
\description{
This function writes an SPM configuration file. Where each element is a command and subcommand from the configuration file
}
\author{
Alistair Dunn
}
