% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbar-kk.R
\name{xbarkk}
\alias{xbarkk}
\title{X-bar chart for known mean and standard deviation of process data}
\usage{
xbarkk(x, mu, sigma, k)
}
\arguments{
\item{x}{is a vector of values representing process data}

\item{mu}{is the process mean}

\item{sigma}{is the process standard deviation}

\item{k}{is the sigma limits for the control chart}
}
\value{
A X-bar control chart
}
\description{
X-bar chart for known mean and standard deviation of process data
}
\examples{
xbarkk(warpbreaks$breaks, 28, 13, 3)
}

